django/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
djangorestframework
gunicorn
idna
protobuf
python-dateutil
pytz
requests
six
sqlparse
urllib3
django-cors-headers
networkx
haversine
numpy
psycopg